フルスクリーンへの移行機能をMenuに移植
